% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimedia.R
\name{format_duration}
\alias{format_duration}
\title{Format Duration}
\usage{
format_duration(seconds)
}
\arguments{
\item{seconds}{Duration in seconds}
}
\value{
Formatted string
}
\description{
Format Duration
}
\keyword{internal}
