% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{generate_ui_code}
\alias{generate_ui_code}
\title{Generate UI Code from Questions}
\usage{
generate_ui_code(questions)
}
\arguments{
\item{questions}{List of question definitions}
}
\value{
Character vector of R code
}
\description{
Generate UI Code from Questions
}
\keyword{internal}
