% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{nf_add_translations}
\alias{nf_add_translations}
\title{Load Translations from List}
\usage{
nf_add_translations(translations, lang)
}
\arguments{
\item{translations}{Named list of translations}

\item{lang}{Language code}
}
\value{
The \code{translations} list, invisibly (the value just stored
  in the package's internal translation table).
}
\description{
Load Translations from List
}
\examples{
nf_add_translations(list(greeting = "Jambo"), "sw")
nf_t("greeting", lang = "sw")
}
