% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_apply_skip_logic}
\alias{nf_apply_skip_logic}
\title{Apply Skip Logic to Question Set}
\usage{
nf_apply_skip_logic(questions, responses)
}
\arguments{
\item{questions}{List of question definitions with skip_logic rules}

\item{responses}{Current responses}
}
\value{
Named list of question_id -> should_show (logical)
}
\description{
Applies skip logic rules to determine which questions should be shown
}
\examples{
questions <- list(
  age = list(),
  drink = list(skip_logic = nf_skip_rule(show_if = "age >= 18"))
)
nf_apply_skip_logic(questions, list(age = 25))
}
