% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{nf_export_csv}
\alias{nf_export_csv}
\title{Export to CSV}
\usage{
nf_export_csv(data, filename = "survey_export.csv", include_metadata = TRUE)
}
\arguments{
\item{data}{Data frame of survey responses}

\item{filename}{Output filename}

\item{include_metadata}{Include metadata columns}
}
\value{
Path to created file
}
\description{
Exports survey data to CSV format
}
\examples{
data <- data.frame(id = 1:3, age = c(25, 40, 31), county = c("Turkana", "Wajir", "Marsabit"))
path <- file.path(tempdir(), "survey_export.csv")
nf_export_csv(data, path)
unlink(path)
}
