% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{nf_export_excel}
\alias{nf_export_excel}
\title{Export to Excel}
\usage{
nf_export_excel(data, filename = "survey_export.xlsx", include_metadata = TRUE)
}
\arguments{
\item{data}{Data frame of survey responses}

\item{filename}{Output filename}

\item{include_metadata}{Include metadata columns}
}
\value{
Path to created file
}
\description{
Exports survey data to Excel format (.xlsx)
}
\examples{
if (requireNamespace("openxlsx", quietly = TRUE)) {
  data <- data.frame(id = 1:3, age = c(25, 40, 31))
  path <- file.path(tempdir(), "survey_export.xlsx")
  nf_export_excel(data, path)
  unlink(path)
}
}
