% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{nf_export_json}
\alias{nf_export_json}
\title{Export to JSON}
\usage{
nf_export_json(data, filename = "survey_export.json", pretty = TRUE)
}
\arguments{
\item{data}{Data frame of survey responses}

\item{filename}{Output filename}

\item{pretty}{Pretty print JSON (default: TRUE)}
}
\value{
Path to created file
}
\description{
Exports survey data to JSON format
}
\examples{
data <- data.frame(id = 1:3, age = c(25, 40, 31))
path <- file.path(tempdir(), "survey_export.json")
nf_export_json(data, path)
unlink(path)
}
