% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{nf_export_rds}
\alias{nf_export_rds}
\title{Export to R Data (.rds)}
\usage{
nf_export_rds(data, filename = "survey_export.rds", compress = "xz")
}
\arguments{
\item{data}{Data frame of survey responses}

\item{filename}{Output filename}

\item{compress}{Compression method (default: "xz")}
}
\value{
Path to created file
}
\description{
Exports survey data to R's native format
}
\examples{
data <- data.frame(id = 1:3, age = c(25, 40, 31))
path <- file.path(tempdir(), "survey_export.rds")
nf_export_rds(data, path)
unlink(path)
}
