% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{nf_export_spss}
\alias{nf_export_spss}
\title{Export to SPSS (.sav)}
\usage{
nf_export_spss(data, filename = "survey_export.sav")
}
\arguments{
\item{data}{Data frame of survey responses}

\item{filename}{Output filename}
}
\value{
Path to created file
}
\description{
Exports survey data to SPSS format
}
\examples{
if (requireNamespace("haven", quietly = TRUE)) {
  df <- data.frame(id = 1:2, score = c(4.5, 3.2))
  path <- file.path(tempdir(), "example.sav")
  nf_export_spss(df, path)
  unlink(path)
}
}
