% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{nf_get_language}
\alias{nf_get_language}
\title{Get Current Language}
\usage{
nf_get_language()
}
\value{
The current language code as a character string.
}
\description{
Get Current Language
}
\examples{
nf_get_language()
}
