% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{nf_initials}
\alias{nf_initials}
\title{Initials Capture}
\usage{
nf_initials(id, label, required = FALSE)
}
\arguments{
\item{id}{Question ID}

\item{label}{Question label}

\item{required}{Is the question required?}
}
\value{
Shiny input
}
\description{
Smaller signature pad for capturing initials
}
\examples{
nf_initials("consent_initials", "Initial each page")
}
