% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{nf_set_language}
\alias{nf_set_language}
\title{Set Current Language}
\usage{
nf_set_language(lang)
}
\arguments{
\item{lang}{Language code (e.g., "en", "es", "fr", "sw")}
}
\value{
The language code, invisibly.
}
\description{
Set Current Language
}
\examples{
old_lang <- nf_get_language()
nf_set_language("es")
nf_get_language()
nf_set_language(old_lang)
}
