% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_show_if_all}
\alias{nf_show_if_all}
\title{Show if Multiple Conditions Met (AND)}
\usage{
nf_show_if_all(...)
}
\arguments{
\item{...}{Skip logic functions or expressions}
}
\value{
Skip logic function
}
\description{
Show if Multiple Conditions Met (AND)
}
\examples{
show_q <- nf_show_if_all(
  nf_show_if_greater("age", 17),
  nf_show_if_equals("consent", TRUE)
)
show_q(list(age = 25, consent = TRUE))
}
