% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_show_if_any}
\alias{nf_show_if_any}
\title{Show if Any Condition Met (OR)}
\usage{
nf_show_if_any(...)
}
\arguments{
\item{...}{Skip logic functions or expressions}
}
\value{
Skip logic function
}
\description{
Show if Any Condition Met (OR)
}
\examples{
show_q <- nf_show_if_any(
  nf_show_if_equals("country", "KE"),
  nf_show_if_equals("country", "UG")
)
show_q(list(country = "UG"))
}
