% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_show_if_contains}
\alias{nf_show_if_contains}
\title{Show if Previous Answer Contains}
\usage{
nf_show_if_contains(previous_question_id, value)
}
\arguments{
\item{previous_question_id}{ID of previous question}

\item{value}{Value to check for}
}
\value{
Skip logic function
}
\description{
For checkbox/multi-select questions
}
\examples{
show_other <- nf_show_if_contains("symptoms", "other")
show_other(list(symptoms = c("fever", "other")))
}
