% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_show_if_equals}
\alias{nf_show_if_equals}
\title{Show if Previous Answer Equals}
\usage{
nf_show_if_equals(previous_question_id, value)
}
\arguments{
\item{previous_question_id}{ID of previous question}

\item{value}{Value to compare against}
}
\value{
Skip logic function
}
\description{
Show if Previous Answer Equals
}
\examples{
show_drink <- nf_show_if_equals("age_group", "adult")
show_drink(list(age_group = "adult"))
}
