% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_show_if_greater}
\alias{nf_show_if_greater}
\title{Show if Previous Answer Greater Than}
\usage{
nf_show_if_greater(previous_question_id, threshold)
}
\arguments{
\item{previous_question_id}{ID of previous question}

\item{threshold}{Threshold value}
}
\value{
Skip logic function
}
\description{
Show if Previous Answer Greater Than
}
\examples{
show_adult_q <- nf_show_if_greater("age", 17)
show_adult_q(list(age = 25))
}
