% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_show_if_in}
\alias{nf_show_if_in}
\title{Show if Previous Answer In List}
\usage{
nf_show_if_in(previous_question_id, values)
}
\arguments{
\item{previous_question_id}{ID of previous question}

\item{values}{Vector of values to check}
}
\value{
Skip logic function
}
\description{
Show if Previous Answer In List
}
\examples{
show_followup <- nf_show_if_in("country", c("KE", "UG", "TZ"))
show_followup(list(country = "KE"))
}
