% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_show_if_less}
\alias{nf_show_if_less}
\title{Show if Previous Answer Less Than}
\usage{
nf_show_if_less(previous_question_id, threshold)
}
\arguments{
\item{previous_question_id}{ID of previous question}

\item{threshold}{Threshold value}
}
\value{
Skip logic function
}
\description{
Show if Previous Answer Less Than
}
\examples{
show_minor_q <- nf_show_if_less("age", 18)
show_minor_q(list(age = 12))
}
