% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_logic.R
\name{nf_skip_logic}
\alias{nf_skip_logic}
\title{Evaluate Skip Logic}
\usage{
nf_skip_logic(condition, responses, default = TRUE)
}
\arguments{
\item{condition}{Expression or function to evaluate}

\item{responses}{Named list of previous responses}

\item{default}{Default value if evaluation fails (default: TRUE)}
}
\value{
Logical indicating whether to show the question
}
\description{
Determines whether a question should be shown based on previous responses
}
\examples{
responses <- list(age = 25)
nf_skip_logic("age >= 18", responses)
nf_skip_logic(function(r) r$age >= 18, responses)
}
