% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{nf_validate_date}
\alias{nf_validate_date}
\title{Validate Date}
\usage{
nf_validate_date(date)
}
\arguments{
\item{date}{Date string}
}
\value{
Logical
}
\description{
Validate Date
}
\examples{
nf_validate_date("2026-07-26")
nf_validate_date("not-a-date")
}
