% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{nf_validate_email}
\alias{nf_validate_email}
\title{Validate Email}
\usage{
nf_validate_email(email)
}
\arguments{
\item{email}{Email address}
}
\value{
Logical
}
\description{
Validate Email
}
\examples{
nf_validate_email("user@example.org")
nf_validate_email("not-an-email")
}
