% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{nf_validate_phone}
\alias{nf_validate_phone}
\title{Validate Phone Number}
\usage{
nf_validate_phone(phone)
}
\arguments{
\item{phone}{Phone number string}
}
\value{
Logical
}
\description{
Validate Phone Number
}
\examples{
nf_validate_phone("+254712345678")
nf_validate_phone("abc")
}
