% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{nf_validate_url}
\alias{nf_validate_url}
\title{Validate URL}
\usage{
nf_validate_url(url)
}
\arguments{
\item{url}{URL string}
}
\value{
Logical
}
\description{
Validate URL
}
\examples{
nf_validate_url("https://example.org/survey")
nf_validate_url("not a url")
}
