% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{normalize_field_type}
\alias{normalize_field_type}
\title{Normalize Field Type}
\usage{
normalize_field_type(type)
}
\arguments{
\item{type}{Field type string}
}
\value{
Normalized type
}
\description{
Normalize Field Type
}
\keyword{internal}
