% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{parse_choices}
\alias{parse_choices}
\title{Parse Generic Choices String}
\usage{
parse_choices(choices)
}
\arguments{
\item{choices}{Comma-separated or pipe-separated choices}
}
\value{
Character vector of choices
}
\description{
Parse Generic Choices String
}
\keyword{internal}
