% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{parse_redcap_branching}
\alias{parse_redcap_branching}
\title{Parse REDCap Branching Logic}
\usage{
parse_redcap_branching(logic)
}
\arguments{
\item{logic}{REDCap branching logic string}
}
\value{
Skip logic expression
}
\description{
Parse REDCap Branching Logic
}
\keyword{internal}
