% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{parse_redcap_choices}
\alias{parse_redcap_choices}
\title{Parse REDCap Choices}
\usage{
parse_redcap_choices(choices)
}
\arguments{
\item{choices}{REDCap choices string (e.g., "1, Yes | 0, No")}
}
\value{
Named character vector of choices
}
\description{
Parse REDCap Choices
}
\keyword{internal}
