% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{parse_redcap_type}
\alias{parse_redcap_type}
\title{Parse REDCap Field Type}
\usage{
parse_redcap_type(type)
}
\arguments{
\item{type}{REDCap field type}
}
\value{
Normalized field type
}
\description{
Parse REDCap Field Type
}
\keyword{internal}
