% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{parse_redcap_validation}
\alias{parse_redcap_validation}
\title{Parse REDCap Validation}
\usage{
parse_redcap_validation(validation)
}
\arguments{
\item{validation}{REDCap validation string}
}
\value{
Validation rules list
}
\description{
Parse REDCap Validation
}
\keyword{internal}
